// src/app/api/model/search-customer-request.ts
/**
 * CrediFlow.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SearchCustomerRequest { 
    keyword?: string | null;
    pageIndex?: number;
    pageSize?: number;
    sortBy?: string | null;
    sortDesc?: boolean;
    hasBadDebt?: boolean | null;
    hasActiveLoan?: boolean | null;
}